+++ .gitignore
@@ -1,7 +1,6 @@
 target/
-!.mvn/wrapper/maven-wrapper.jar
-!**/src/main/**/target/
-!**/src/test/**/target/
+sample/out/*
+!sample/out/.gitkeep
 
 ### IntelliJ IDEA ###
 .idea/
@@ -11,3 +10,4 @@ target/
 
 ### Mac OS ###
 .DS_Store
+
