+++ src/main/java/module-info.java
@@ -8,5 +8,5 @@ module app {
     requires javafx.base;
     requires javafx.controls;
     requires javafx.graphics;
-    requires opencv;
+    requires transitive opencv;
 }
